requirements.txt
coverage >= 7.3.2
pandas >= 2.1.3
python-dotenv >= 1.0.0